+++ modules/casino.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require_relative '../modules/main'
 
 class Casino < Main
