+++ modules/main.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require_relative '../interface/view'
 
 class Main
@@ -6,10 +8,17 @@ class Main
 
   @@bank = 0
   @@games = 0
-  @@stack = { '2_ЧЕРВИ' => 2, '3_ЧЕРВИ' => 3, '4_ЧЕРВИ' => 4, '5_ЧЕРВИ' => 5, '6_ЧЕРВИ' => 6, '7_ЧЕРВИ' => 7, '8_ЧЕРВИ' => 8, '9_ЧЕРВИ' => 9, '10_ЧЕРВИ' => 10, 'V_ЧЕРВИ' => 2, 'Q_ЧЕРВИ' => 3, 'K_ЧЕРВИ' => 4, 'A_ЧЕРВИ' => 11,
-              '2_БУБНЫ' => 2, '3_БУБНЫ' => 3, '4_БУБНЫ' => 4, '5_БУБНЫ' => 5, '6_БУБНЫ' => 6, '7_БУБНЫ' => 7, '8_БУБНЫ' => 8, '9_БУБНЫ' => 9, '10_БУБНЫ' => 10, 'V_БУБНЫ' => 2, 'Q_БУБНЫ' => 3, 'K_БУБНЫ' => 4, 'A_БУБНЫ' => 11,
-              '2_ПИКИ' => 2, '3_ПИКИ' => 3, '4_ПИКИ' => 4, '5_ПИКИ' => 5, '6_ПИКИ' => 6, '7_ПИКИ' => 7, '8_ПИКИ' => 8, '9_ПИКИ' => 9, '10_ПИКИ' => 10, 'V_ПИКИ' => 2, 'Q_ПИКИ' => 3, 'K_ПИКИ' => 4, 'A_ПИКИ' => 11,
-              '2_КРЕСТЫ' => 2, '3_КРЕСТЫ' => 3, '4_КРЕСТЫ' => 4, '5_КРЕСТЫ' => 5, '6_КРЕСТЫ' => 6, '7_КРЕСТЫ' => 7, '8_КРЕСТЫ' => 8, '9_КРЕСТЫ' => 9, '10_КРЕСТЫ' => 10, 'V_КРЕСТЫ' => 2, 'Q_КРЕСТЫ' => 3, 'K_КРЕСТЫ' => 4, 'A_КРЕСТЫ' => 11 }
+  @@stack = { '2_ЧЕРВИ' => 2, '3_ЧЕРВИ' => 3, '4_ЧЕРВИ' => 4, '5_ЧЕРВИ' => 5, '6_ЧЕРВИ' => 6,
+              '7_ЧЕРВИ' => 7, '8_ЧЕРВИ' => 8, '9_ЧЕРВИ' => 9, '10_ЧЕРВИ' => 10, 'V_ЧЕРВИ' => 2,
+              'Q_ЧЕРВИ' => 3, 'K_ЧЕРВИ' => 4, 'A_ЧЕРВИ' => 11, '2_БУБНЫ' => 2, '3_БУБНЫ' => 3,
+              '4_БУБНЫ' => 4, '5_БУБНЫ' => 5, '6_БУБНЫ' => 6, '7_БУБНЫ' => 7, '8_БУБНЫ' => 8,
+              '9_БУБНЫ' => 9, '10_БУБНЫ' => 10, 'V_БУБНЫ' => 2, 'Q_БУБНЫ' => 3, 'K_БУБНЫ' => 4,
+              'A_БУБНЫ' => 11, '2_ПИКИ' => 2, '3_ПИКИ' => 3, '4_ПИКИ' => 4, '5_ПИКИ' => 5,
+              '6_ПИКИ' => 6, '7_ПИКИ' => 7, '8_ПИКИ' => 8, '9_ПИКИ' => 9, '10_ПИКИ' => 10,
+              'V_ПИКИ' => 2, 'Q_ПИКИ' => 3, 'K_ПИКИ' => 4, 'A_ПИКИ' => 11, '2_КРЕСТЫ' => 2,
+              '3_КРЕСТЫ' => 3, '4_КРЕСТЫ' => 4, '5_КРЕСТЫ' => 5, '6_КРЕСТЫ' => 6, '7_КРЕСТЫ' => 7,
+              '8_КРЕСТЫ' => 8, '9_КРЕСТЫ' => 9, '10_КРЕСТЫ' => 10, 'V_КРЕСТЫ' => 2, 'Q_КРЕСТЫ' => 3,
+              'K_КРЕСТЫ' => 4, 'A_КРЕСТЫ' => 11 }
 
   def self.games_count
     @@games += 1
@@ -49,9 +58,9 @@ class Main
     View.bank
     puts "#{@@bank} $"
     View.line
-    print "Вы(очки: #{player.points}): "
+    print "Вы(очки: " + "#{player.points}" + "):" 
     player.show_hand
-    print "Дилер(очки: #{casino.points}): "
+    print "Дилер(очки: " + "#{casino.points}" + "):" 
     casino.show_hand
     View.line
   end
@@ -90,9 +99,16 @@ class Main
     casino.hand = []
     player.points = 0
     casino.points = 0
-    @@stack = { '2_ЧЕРВИ' => 2, '3_ЧЕРВИ' => 3, '4_ЧЕРВИ' => 4, '5_ЧЕРВИ' => 5, '6_ЧЕРВИ' => 6, '7_ЧЕРВИ' => 7, '8_ЧЕРВИ' => 8, '9_ЧЕРВИ' => 9, '10_ЧЕРВИ' => 10, 'V_ЧЕРВИ' => 2, 'Q_ЧЕРВИ' => 3, 'K_ЧЕРВИ' => 4, 'A_ЧЕРВИ' => 11,
-                '2_БУБНЫ' => 2, '3_БУБНЫ' => 3, '4_БУБНЫ' => 4, '5_БУБНЫ' => 5, '6_БУБНЫ' => 6, '7_БУБНЫ' => 7, '8_БУБНЫ' => 8, '9_БУБНЫ' => 9, '10_БУБНЫ' => 10, 'V_БУБНЫ' => 2, 'Q_БУБНЫ' => 3, 'K_БУБНЫ' => 4, 'A_БУБНЫ' => 11,
-                '2_ПИКИ' => 2, '3_ПИКИ' => 3, '4_ПИКИ' => 4, '5_ПИКИ' => 5, '6_ПИКИ' => 6, '7_ПИКИ' => 7, '8_ПИКИ' => 8, '9_ПИКИ' => 9, '10_ПИКИ' => 10, 'V_ПИКИ' => 2, 'Q_ПИКИ' => 3, 'K_ПИКИ' => 4, 'A_ПИКИ' => 11,
-                '2_КРЕСТЫ' => 2, '3_КРЕСТЫ' => 3, '4_КРЕСТЫ' => 4, '5_КРЕСТЫ' => 5, '6_КРЕСТЫ' => 6, '7_КРЕСТЫ' => 7, '8_КРЕСТЫ' => 8, '9_КРЕСТЫ' => 9, '10_КРЕСТЫ' => 10, 'V_КРЕСТЫ' => 2, 'Q_КРЕСТЫ' => 3, 'K_КРЕСТЫ' => 4, 'A_КРЕСТЫ' => 11 }
+    @@stack = { '2_ЧЕРВИ' => 2, '3_ЧЕРВИ' => 3, '4_ЧЕРВИ' => 4, '5_ЧЕРВИ' => 5, '6_ЧЕРВИ' => 6,
+                '7_ЧЕРВИ' => 7, '8_ЧЕРВИ' => 8, '9_ЧЕРВИ' => 9, '10_ЧЕРВИ' => 10, 'V_ЧЕРВИ' => 2,
+                'Q_ЧЕРВИ' => 3, 'K_ЧЕРВИ' => 4, 'A_ЧЕРВИ' => 11, '2_БУБНЫ' => 2, '3_БУБНЫ' => 3,
+                '4_БУБНЫ' => 4, '5_БУБНЫ' => 5, '6_БУБНЫ' => 6, '7_БУБНЫ' => 7, '8_БУБНЫ' => 8,
+                '9_БУБНЫ' => 9, '10_БУБНЫ' => 10, 'V_БУБНЫ' => 2, 'Q_БУБНЫ' => 3, 'K_БУБНЫ' => 4,
+                'A_БУБНЫ' => 11, '2_ПИКИ' => 2, '3_ПИКИ' => 3, '4_ПИКИ' => 4, '5_ПИКИ' => 5,
+                '6_ПИКИ' => 6, '7_ПИКИ' => 7, '8_ПИКИ' => 8, '9_ПИКИ' => 9, '10_ПИКИ' => 10,
+                'V_ПИКИ' => 2, 'Q_ПИКИ' => 3, 'K_ПИКИ' => 4, 'A_ПИКИ' => 11, '2_КРЕСТЫ' => 2,
+                '3_КРЕСТЫ' => 3, '4_КРЕСТЫ' => 4, '5_КРЕСТЫ' => 5, '6_КРЕСТЫ' => 6, '7_КРЕСТЫ' => 7,
+                '8_КРЕСТЫ' => 8, '9_КРЕСТЫ' => 9, '10_КРЕСТЫ' => 10, 'V_КРЕСТЫ' => 2, 'Q_КРЕСТЫ' => 3,
+                'K_КРЕСТЫ' => 4, 'A_КРЕСТЫ' => 11 }
   end
 end
