+++ modules/player.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require_relative '../modules/main'
 
 class Player < Main
